change over to using usize everywhere. after thinking, probably best

+++ src/lib.rs
@@ -3,36 +3,36 @@ pub mod hashlrurs {
 
     #[derive(Clone)]
     pub struct HashLRU {
-        size: u128,
-        max: u128,
-        // NOTE: Currently only supports u128 due to to it only being used
+        size: usize,
+        max: usize,
+        // NOTE: Currently only supports usize due to to it only being used
         //          in hyperswarm to hold integeters here:
         // https://github.com/hyperswarm/dht/blob/c8bbe643dac374d9c7cf92d723b732a980c564bd/index.js#L36
-        cache1: HashMap<u128, u128>,
-        cache2: HashMap<u128, u128>,
+        cache1: HashMap<usize, usize>,
+        cache2: HashMap<usize, usize>,
         state: bool,
     }
 
     impl HashLRU {
-        pub fn new(&self, max: u128) -> HashLRU {
+        pub fn new(&self, max: usize) -> HashLRU {
             HashLRU {
                 size : 0,
                 max,
-                cache1: HashMap::with_capacity(max as usize),
-                cache2: HashMap::with_capacity(max as usize),
+                cache1: HashMap::with_capacity(max),
+                cache2: HashMap::with_capacity(max),
                 state: true,
             }
         }
-        pub fn has(&self, key: u128) -> bool {
+        pub fn has(&self, key: usize) -> bool {
             self.cache1.contains_key(&key) || self.cache2.contains_key(&key)
         }
 
-        pub fn remove(&mut self, key: u128) {
+        pub fn remove(&mut self, key: usize) {
             self.cache1.remove(&key).unwrap();
             self.cache2.remove(&key).unwrap();
         }
 
-        pub fn get(&mut self, key: u128) -> Option<u128> {
+        pub fn get(&mut self, key: usize) -> Option<usize> {
             if self.state {
                 match self.cache1.get(&key) {
                     Some(&v) => return Some(v),
@@ -62,7 +62,7 @@ pub mod hashlrurs {
             }
         }
 
-        pub fn set(&mut self, key: u128, value: u128) {
+        pub fn set(&mut self, key: usize, value: usize) {
             if self.state {
             match self.cache1.get(&key) {
                 Some(_v) => {self.cache1.insert(key, value);},
@@ -76,7 +76,7 @@ pub mod hashlrurs {
         }
         }
 
-        pub fn update(&mut self, key: u128, value: u128) {
+        pub fn update(&mut self, key: usize, value: usize) {
             if self.state {
                 self.cache1.insert(key, value);
                 self.size += 1;
